index.php
This little bit of PHP will load a rss feed from a wordpress site.  It will pick up the first story and convert it to html, then it will display the next 5 links and create links for you.  This works well on sites where you don't want to edit the html such as an index.php file so you can edit the front page of your site using wordpress and have the results show up immediately.